[Nextjs] Make `extractPath` shared and remove duplicate `page-props-factory/plugins/normalMode` for `personalize` addon
