+++ packages/create-sitecore-jss/src/templates/nextjs/src/lib/page-props-factory/extractPath.ts
@@ -1,0 +1,19 @@
+import { ParsedUrlQuery } from 'querystring';
+
+/**
+ * Extract normalized Sitecore item path from query
+ * @param {ParsedUrlQuery | undefined} params
+ */
+export function extractPath(params: ParsedUrlQuery | undefined): string {
+  if (params === undefined) {
+    return '/';
+  }
+  let path = Array.isArray(params.path) ? params.path.join('/') : params.path ?? '/';
+
+  // Ensure leading '/'
+  if (!path.startsWith('/')) {
+    path = '/' + path;
+  }
+
+  return path;
+}
